Last run

See merge request phil-all/snowtricks!11

+++ src/Controller/TrickController.php
@@ -29,7 +29,7 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 class TrickController extends AbstractController
 {
     /**
-     * @Route("/", name="app_home")
+     * @Route("/", name="app_home", methods={"GET"})
      *
      * @param TrickRepository $trickRepository
      *
@@ -43,7 +43,7 @@ class TrickController extends AbstractController
     }
 
     /**
-     * @Route("/trick/{id}/{slug}", name="app_trick_read")
+     * @Route("/trick/{id}/{slug}", name="app_trick_read", methods={"GET", "POST"})
      *
      * @param Request           $request
      * @param Trick             $trick
@@ -79,12 +79,14 @@ class TrickController extends AbstractController
     }
 
     /**
-     * @Route("/trick/modifier/{id}/{slug}", name="app_trick_update")
-     * @Route("/create", name="app_trick_create")
+     * @Route("/trick/modifier/{id}/{slug}", name="app_trick_update", methods={"GET", "POST"})
+     * @Route("/create", name="app_trick_create", methods={"GET", "POST"})
      *
-     * @param Request              $request
-     * @param CategoryRepository   $categoryRepository
-     * @param TrickRepository      $trickRepository
+     * @param Request             $request
+     * @param TrickService        $trickService
+     * @param TrickRepository     $trickRepository
+     * @param CategoryRepository  $categoryRepository
+     * @param MediaUpdaterService $mediaUpdaterService
      *
      * @return Response
      */
@@ -146,7 +148,7 @@ class TrickController extends AbstractController
                 ->defineTrick($trick)
                 ->updateMedias($uploadedThumbnailFile, $images, $videos);
 
-            $trickRepository->update($trick);
+            $trickService->update($trick);
 
             return $this->redirectToRoute('app_home');
         }
@@ -158,16 +160,31 @@ class TrickController extends AbstractController
     }
 
     /**
-     * @Route("/trick/delete/{id}/{slug}", name="app_trick_delete")
+     * @Route("/trick/delete/{id}/{token}", name="app_trick_delete", methods={"GET"})
+     *
+     * @param Trick        $trick
+     * @param Request      $request
+     * @param TrickService $trickService
      *
      * @return RedirectResponse
      */
-    public function delete(Trick $trick, TrickService $trickService): RedirectResponse
-    {
+    public function delete(
+        Trick $trick,
+        Request $request,
+        TrickService $trickService
+    ): RedirectResponse {
+
         $this->denyAccessUnlessGranted('ROLE_USER');
 
-        $trickService->delete($trick);
+        /** @var string $submittedToken */
+        $submittedToken = $request->attributes->get('_route_params')['token'];
 
-        return $this->redirectToRoute('app_home');
+        if ($this->isCsrfTokenValid('delete-trick', $submittedToken)) {
+            $trickService->delete($trick);
+
+            return $this->redirectToRoute('app_home');
+        }
+
+        return $this->redirectToRoute('app_error_visitor_link');
     }
 }
